+++ Sources/Subprocess/IO/LockedState.swift
@@ -12,7 +12,7 @@
 
 #if canImport(os)
 internal import os
-#if FOUNDATION_FRAMEWORK && canImport(C.os.lock)
+#if canImport(C.os.lock)
 internal import C.os.lock
 #endif
 #elseif canImport(Bionic)
